requirements.txt
appdirs==1.4.3
attrs==19.3.0
beautifulsoup4==4.8.2
black==19.10b0
bs4==0.0.1
cached-property==1.5.1
certifi==2019.11.28
chardet==3.0.4
click==7.1.1
idna==2.9
pathspec==0.7.0
pyaml==20.3.1
PyYAML==5.3.1
regex==2020.2.20
requests==2.23.0
soupsieve==2.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.8
